refactor: 주 도메인 패키지 구조 재설정
* 주문, 주문테이블 주문 도메인으로 묶고 메뉴 도메인과 패키지 단방향 관계
