--- src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -1,58 +1,0 @@
-package kitchenpos.order.domain;
-
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-
-import kitchenpos.menu.domain.Menu;
-
-@Entity
-public class OrderLineItem {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long seq;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Order order;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Menu menu;
-
-    private long quantity;
-
-    protected OrderLineItem() {
-    }
-
-    public OrderLineItem(final Menu menu, final long quantity) {
-        this.menu = menu;
-        this.quantity = quantity;
-    }
-
-    public Long getSeq() {
-        return seq;
-    }
-
-    public Order getOrder() {
-        return order;
-    }
-
-    public Menu getMenu() {
-        return menu;
-    }
-
-    public long getQuantity() {
-        return quantity;
-    }
-
-    public Long getMenuId() {
-        return this.menu.getId();
-    }
-
-    protected void updateOrder(final Order order) {
-        this.order = order;
-    }
-}
