--- src/main/java/kitchenpos/order/domain/OrderStatus.java
@@ -1,22 +1,0 @@
-package kitchenpos.order.domain;
-
-import java.util.Arrays;
-
-public enum OrderStatus {
-    COOKING, MEAL, COMPLETION;
-
-    public static OrderStatus of(final String requestOrderStatus) {
-        return Arrays.stream(OrderStatus.values())
-                .filter(orderStatus -> requestOrderStatus.equalsIgnoreCase(orderStatus.name()))
-                .findFirst()
-                .orElseThrow(IllegalAccessError::new);
-    }
-
-    public boolean isCookingOrMeal() {
-        return this == COOKING || this == MEAL;
-    }
-
-    public boolean isCompletion() {
-        return this == COMPLETION;
-    }
-}
