+++ src/main/java/kitchenpos/order/domain/order/OrderLineItemRepository.java
@@ -1,0 +1,6 @@
+package kitchenpos.order.domain.order;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
+}
