+++ src/main/java/kitchenpos/order/domain/order/OrderRepository.java
@@ -1,0 +1,8 @@
+package kitchenpos.order.domain.order;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import kitchenpos.order.domain.order.Order;
+
+public interface OrderRepository extends JpaRepository<Order, Long> {
+}
