+++ src/main/java/kitchenpos/order/domain/ordertable/OrderTableRepository.java
@@ -1,0 +1,9 @@
+package kitchenpos.order.domain.ordertable;
+
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
+	List<OrderTable> findAllByIdIn(List<Long> orderTableIds);
+}
