+++ src/main/java/kitchenpos/order/domain/ordertable/TableGroupRepository.java
@@ -1,0 +1,6 @@
+package kitchenpos.order.domain.ordertable;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
+}
