+++ src/main/java/kitchenpos/order/ui/ordertable/TableRestController.java
@@ -1,0 +1,56 @@
+package kitchenpos.order.ui.ordertable;
+
+import kitchenpos.order.application.ordertable.TableService;
+import kitchenpos.order.dto.ordertable.OrderTableRequest;
+import kitchenpos.order.dto.ordertable.OrderTableResponse;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+public class TableRestController {
+    private final TableService tableService;
+
+    public TableRestController(final TableService tableService) {
+        this.tableService = tableService;
+    }
+
+    @PostMapping("/api/tables")
+    public ResponseEntity<OrderTableResponse> create(@RequestBody final OrderTableRequest orderTableRequest) {
+        final OrderTableResponse created = tableService.create(orderTableRequest);
+        final URI uri = URI.create("/api/tables/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created)
+                ;
+    }
+
+    @GetMapping("/api/tables")
+    public ResponseEntity<List<OrderTableResponse>> list() {
+        return ResponseEntity.ok()
+                .body(tableService.findAll())
+                ;
+    }
+
+    @PutMapping("/api/tables/{orderTableId}/empty")
+    public ResponseEntity<OrderTableResponse> changeEmpty(
+            @PathVariable final Long orderTableId,
+            @RequestBody final OrderTableRequest orderTableRequest
+    ) {
+        return ResponseEntity.ok()
+                .body(tableService.changeEmpty(orderTableId, orderTableRequest))
+                ;
+    }
+
+    @PutMapping("/api/tables/{orderTableId}/number-of-guests")
+    public ResponseEntity<OrderTableResponse> changeNumberOfGuests(
+            @PathVariable final Long orderTableId,
+            @RequestBody final OrderTableRequest orderTableRequest
+    ) {
+        return ResponseEntity.ok()
+                .body(tableService.changeNumberOfGuests(orderTableId, orderTableRequest))
+                ;
+    }
+}
