--- src/main/java/kitchenpos/ordertable/application/TableGroupService.java
@@ -1,39 +1,0 @@
-package kitchenpos.ordertable.application;
-
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.OrderTableRepository;
-import kitchenpos.ordertable.domain.TableGroup;
-import kitchenpos.ordertable.domain.TableGroupRepository;
-import kitchenpos.ordertable.dto.TableGroupRequest;
-import kitchenpos.ordertable.dto.TableGroupResponse;
-
-@Transactional
-@Service
-public class TableGroupService {
-    private final OrderTableRepository orderTableRepository;
-    private final TableGroupRepository tableGroupRepository;
-
-    public TableGroupService(final OrderTableRepository orderTableRepository,
-        final TableGroupRepository tableGroupRepository) {
-        this.orderTableRepository = orderTableRepository;
-        this.tableGroupRepository = tableGroupRepository;
-    }
-
-    public TableGroupResponse create(final TableGroupRequest tableGroupRequest) {
-        List<OrderTable> orderTables = orderTableRepository.findAllByIdIn(tableGroupRequest.getOrderTableIds());
-        TableGroup tableGroup = tableGroupRequest.toTableGroup(orderTables);
-
-        return TableGroupResponse.of(tableGroupRepository.save(tableGroup));
-    }
-
-    public void ungroup(final Long tableGroupId) {
-        TableGroup tableGroup = tableGroupRepository.findById(tableGroupId)
-            .orElseThrow(IllegalArgumentException::new);
-        tableGroup.ungroup();
-    }
-}
