--- src/main/java/kitchenpos/ordertable/application/TableService.java
@@ -1,48 +1,0 @@
-package kitchenpos.ordertable.application;
-
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.OrderTableRepository;
-import kitchenpos.ordertable.dto.OrderTableRequest;
-import kitchenpos.ordertable.dto.OrderTableResponse;
-
-@Transactional
-@Service
-public class TableService {
-    private final OrderTableRepository orderTableRepository;
-
-    public TableService(final OrderTableRepository orderTableRepository) {
-        this.orderTableRepository = orderTableRepository;
-    }
-
-    public OrderTableResponse create(final OrderTableRequest orderTableRequest) {
-        OrderTable orderTable = orderTableRequest.toOrderTable();
-        return OrderTableResponse.of(orderTableRepository.save(orderTable));
-    }
-
-    @Transactional(readOnly = true)
-    public List<OrderTableResponse> findAll() {
-        return OrderTableResponse.ofList(orderTableRepository.findAll());
-    }
-
-    public OrderTableResponse changeEmpty(final Long orderTableId, final OrderTableRequest orderTableRequest) {
-        final OrderTable orderTable = findById(orderTableId);
-        orderTable.changeEmpty(orderTableRequest.isEmpty());
-        return OrderTableResponse.of(orderTable);
-    }
-
-    public OrderTableResponse changeNumberOfGuests(final Long orderTableId, final OrderTableRequest orderTableRequest) {
-        final OrderTable orderTable = findById(orderTableId);
-        orderTable.changeNumberOfGuests(orderTableRequest.getNumberOfGuests());
-        return OrderTableResponse.of(orderTable);
-    }
-
-    private OrderTable findById(final Long orderTableId) {
-        return orderTableRepository.findById(orderTableId)
-            .orElseThrow(IllegalArgumentException::new);
-    }
-}
