--- src/main/java/kitchenpos/ordertable/domain/OrderTableRepository.java
@@ -1,9 +1,0 @@
-package kitchenpos.ordertable.domain;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
-	List<OrderTable> findAllByIdIn(List<Long> orderTableIds);
-}
