--- src/main/java/kitchenpos/ordertable/domain/TableGroupRepository.java
@@ -1,6 +1,0 @@
-package kitchenpos.ordertable.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
-}
